first2 commit

+++ database/migrations/2023_05_22_160000_create_foreign_keys.php
@@ -14,9 +14,12 @@ class CreateForeignKeys extends Migration {
 						->onUpdate('cascade');
 		});
 		Schema::table('facilits', function(Blueprint $table) {
+
+			if (Schema::hasColumn('facilits', 'owner_id')) {
 			$table->foreign('town_id')->references('id')->on('towns')
 						->onDelete('cascade')
 						->onUpdate('cascade');
+			}
 		});
 		Schema::table('facilits', function(Blueprint $table) {
 			$table->foreign('owner_id')->references('id')->on('owners')
